strict care formatting size_t // rm(trailing white space)
git-svn-id: https://svn.r-project.org/R/trunk@72926 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -241,10 +241,10 @@ static void NORET set_iconv_error(Rconnection con, char* from, char* to)
 static size_t buff_set_len(Rconnection con, size_t len) {
     size_t unread_len = 0;
     unsigned char *buff;
-    
+
     if (con->buff_len == len)
 	return len;
-    
+
     if (con->buff) {
 	unread_len = con->buff_stored_len - con->buff_pos;
 	len = MAX(len, unread_len);
@@ -256,7 +256,7 @@ static size_t buff_set_len(Rconnection con, size_t len) {
 	memcpy(buff, con->buff + con->buff_pos, unread_len);
 	free(con->buff);
     }
-    
+
     con->buff = buff;
     con->buff_len = len;
     con->buff_pos = 0;
@@ -283,12 +283,12 @@ static void buff_reset(Rconnection con) {
 
 static size_t buff_fill(Rconnection con) {
     size_t free_len, read_len;
-    
+
     buff_reset(con);
 
     free_len = con->buff_len - con->buff_stored_len;
     read_len = con->read(con->buff, sizeof(unsigned char), free_len, con);
-    
+
     con->buff_stored_len += read_len;
 
     return read_len;
@@ -297,7 +297,7 @@ static size_t buff_fill(Rconnection con) {
 static int buff_fgetc(Rconnection con)
 {
     size_t unread_len;
-    
+
     unread_len = con->buff_stored_len - con->buff_pos;
     if (unread_len == 0) {
 	size_t filled_len = buff_fill(con);
@@ -314,7 +314,7 @@ static double buff_seek(Rconnection con, double where, int origin, int rw)
 
     if (rw == 2) /* write */
 	return con->seek(con, where, origin, rw);
-    
+
     if (ISNA(where)) /* tell */
 	return con->seek(con, where, origin, rw) - unread_len;
 
@@ -327,11 +327,11 @@ static double buff_seek(Rconnection con, double where, int origin, int rw)
 	}
     }
     con->buff_pos = con->buff_stored_len = 0;
-    
+
     return con->seek(con, where, origin, rw);
 }
 
-void set_buffer(Rconnection con) {    
+void set_buffer(Rconnection con) {
     if (con->canread && con->text) {
 	buff_init(con);
     }
@@ -676,9 +676,9 @@ static Rboolean file_open(Rconnection con)
 	char mode[20]; /* 4 byte mode plus "t,ccs=UTF-16LE" plus one for luck. */
 	strncpy(mode, con->mode, 4);
 	mode[4] = '\0';
-	if (!strpbrk(mode, "bt")) 
+	if (!strpbrk(mode, "bt"))
 	    strcat(mode, "t");
-	if (strchr(mode, 't') 
+	if (strchr(mode, 't')
 	    && (!strcmp(con->encname, "UTF-16LE") || !strcmp(con->encname, "UCS-2LE"))) {
 	    strcat(mode, ",ccs=UTF-16LE");
 	    if (con->canread) {
@@ -807,7 +807,7 @@ static int file_fgetc_internal(Rconnection con)
     	    this->have_wcbuffered = TRUE;
     	}
     } else
-#endif  
+#endif
     c =fgetc(fp);
     return feof(fp) ? R_EOF : c;
 }
@@ -3709,7 +3709,8 @@ size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize)
     size_t nbuf = -1;
 
     while((c = Rconn_fgetc(con)) != R_EOF) {
-	if(nbuf+1 >= bufsize) error(_("line longer than buffer size %d"), bufsize);
+	if(nbuf+1 >= bufsize)
+	    error(_("line longer than buffer size %lu"), (unsigned long) bufsize);
 	if(c != '\n'){
 	    buf[++nbuf] = (char) c;
 	} else {
@@ -3721,7 +3722,8 @@ size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize)
      *  file did not end with newline.
      */
     if(nbuf >= 0 && buf[nbuf]) {
-	if(nbuf+1 >= bufsize) error(_("line longer than buffer size %d"), bufsize);
+	if(nbuf+1 >= bufsize)
+	    error(_("line longer than buffer size %lu"), (unsigned long) bufsize);
 	buf[++nbuf] = '\0';
     }
     return(nbuf);
@@ -4866,7 +4868,7 @@ SEXP attribute_hidden do_writechar(SEXP call, SEXP op, SEXP args, SEXP env)
 	}
     }
     if(!wasopen) {
-        endcontext(&cntxt); 
+        endcontext(&cntxt);
         checkClose(con);
     }
     if(isRaw) {
@@ -5031,7 +5033,7 @@ switch_or_tee_stdout(int icon, int closeOnExit, int tee)
 		Rconnection con = getConnection(icon);
 		R_ReleaseObject(con->ex_ptr);
 		if(SinkConsClose[R_SinkNumber + 1] == 1) { /* close it */
-		    checkClose(con);    
+		    checkClose(con);
 		} else if (SinkConsClose[R_SinkNumber + 1] == 2) /* destroy it */
 		    con_destroy(icon);
 	    }
